Ft bigint bkp

+++ src/ft_math/ft_abs.c
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_abs.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alamit <alamit@student.42.us.org>          +#+  +:+       +#+        */
+/*   By: alamit <alamit@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/25 15:53:48 by alamit            #+#    #+#             */
-/*   Updated: 2019/01/25 15:56:00 by alamit           ###   ########.fr       */
+/*   Updated: 2019/03/28 14:13:39 by alamit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_abs(int n)
+#include <ft_math.h>
+
+uint64_t	ft_abs(int64_t n)
 {
 	if (n < 0)
-		return ((unsigned int)-n);
-	return ((unsigned int)n);
+		return ((uint64_t)-n);
+	return ((uint64_t)n);
 }
